+++ blueprints/incredible-squaring-eigenlayer/Cargo.toml
@@ -58,6 +58,7 @@ num-bigint = { workspace = true }
 [dev-dependencies]
 blueprint-test-utils = { workspace = true }
 gadget-io = { workspace = true }
+tempfile = { workspace = true }
 
 [features]
 default = ["std"]
